Create controller and move some logic over
