+++ Assets/Fungus/Narrative/Scripts/PortraitController.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: cc7c4e379c0d3bb4eb7537ad653ea7f6
+timeCreated: 1465412488
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
